python3unittest: only run UT in native recipes
Summary:
The existing python3unittest.bbclass adds a unit test step after
recipe compile.  It isn't appropriate to run this on target recipes
though because the results of a target recipe can only be reliably ran
on a device.  Set this up so that it only runs on native recipes,
which would use an appropriately configured python running on the build
host.

Test Plan: Build sonorapass and tiogapass.

Reviewed By: benwei13

fbshipit-source-id: 966cbeea61