sonorapass: plat: add full path to fan-util system calls
Summary:
The plat functions to power on and off the server are often called
from other tools, like `fw-util`.  Currently, we use `fan-util` to
force the fan speed when powering on and off the server.  If someone
were to invoke by ssh like `ssh <system> "fw-util ..."`, then the
full profile is not initialized and /usr/local/bin is not in the PATH.

Modify the system calls to have the full path to `fan-util` so that
it is found no matter the PATH state.

Test Plan: Build sonorapass.

Reviewed By: benwei13

fbshipit-source-id: 56f12613fa