fby3: libbic: rename the declaritioin of CMD_OEM_GET_BOARD_ID (#1035)
Summary:
CMD_OEM_GET_BOARD_ID(0Ah) is defined in bic_xfer.h and it is conflict with the declaritioin of CMD_OEM_GET_BOARD_ID(37h) in ipmi.h.
Because one of the declaritioin of CMD_OEM_GET_BOARD_ID is declared in enumerator and the other is declared in macro.
Therefore, the image will be built successfully but the declaration in ipmi.h will be overwritten by the declaration in bic_xfer.h and result in the command(37h) failed.

Declare those BIC commands in enumerator and rename them to distinguish between BMC and BIC command.
Pull Request resolved: https://github.com/facebookexternal/openbmc.wiwynn/pull/1035

Test Plan:
-build code(pass)
-run `ipmitool raw 0x30 0x37`(pass)
[root@localhost ~]# ipmitool raw 0x30 0x37
 0e 01 02 00

Reviewed By: williamspatrick

fbshipit-source-id: 8b3a31fea0