common: rest-api: use temporary file in unit test (#1022)
Summary:
-When a user run `bitbake rest-api` or `bitbake $image`, the unit test of the rest-api is run.
And a file called rest.log that belongs to the user is generated in /tmp/.
Other users will fail to build the image or rest-api package since they don't have the permission to modify the file.
Only when the user removes it, they can build the image or the package.

Make the unit test use the temporary file instead of rest.log.
And the temporary file will be removed when the unit test is done.

Error log:
```````````
|     return open(self.baseFilename, self.mode, encoding=self.encoding)
| PermissionError: [Errno 13] Permission denied: '/tmp/rest.log'
```````````
Pull Request resolved: https://github.com/facebookexternal/openbmc.wiwynn/pull/1022

Test Plan: -build code(pass)

Reviewed By: amithash, deathowl

fbshipit-source-id: d5a93fa9ec