ptest-meson-crosstarget: add script to fixup meson data
Summary:
In order to run ptest on meson-built packages, we need to be
able to run 'meson test' on the target.  Meson has some data
files that describe how to run the test, but they have
information from the build-host.  Process the meson private
data to enable it to run on the target machine.

Test Plan: Build later commit using ptest.

Reviewed By: benwei13

fbshipit-source-id: 20c4fd6a8b