kv: eat ENOENT errors on kv_get
Summary:
It was reported by one of the ODMs that there are numerous cases
where kv_get is called for entries which are not (and likely never
will be) created.  If we log a warning for this we end up filling
up the journal, so we need to avoid logging warnings in this scenario.

One option was to go back to the "old implementation" which only logs
messages if compiled with DEBUG, but this has potential that we are
missing valid file system errors.

Test Plan: Add unit test.  Run in QEMU.

Reviewed By: benwei13

fbshipit-source-id: 7dd7e9afc5