udev: use parent attribute in data0 rule
Summary:
Previously, we were getting a warning about using a deprecated sysfs
attribute 'device', which was being used to find the MTD name of the
data0 device's mtdblockN.  The 'device' is a deprecated method to get
us to the underlying mtdN device for the mtdblockN.  It turns out that
the mtdN is already a udev-parent of mtdblockN, so we can simply use
`ATTR{name}` to get the name from the parent instead of
`ATTR{device/name}`.

Test Plan:
Booted Sonorapass on QEMU and observed warning gone but
device symlink still present.

Reviewed By: tao-ren

fbshipit-source-id: 0300082839