libkv: eat key-already-exists logs for persistent keys
Summary:
It has been noticed that a common pattern for initializing persistent
keys is for a program to blindly call kv_set(...KV_FCREATE).  The affect
of this is that an already existing key will maintain its current value
and a non-existing key will be initialized with the appropriate values.

Detect this pattern and eat the WARN to syslog to avoid filling up the
syslog with needless entries.

Test Plan:
Added unit test to create this condition and observe the WARN call is no
longer present after code change.

Reviewed By: tao-ren

fbshipit-source-id: 7568d207e4