libpal_0.2: simplify bbappends and depend on obmc-pal
Summary:
All machines are required to provide pal.c and pal.h files already,
so simplify the machine bbappends by putting it into the base recipe
and meson.build.

Previously, I attempted to unravel the dual package structure of
obmc-pal and libpal.  This worked for some of the machines, but in
porting libpal_0.2 to other machines I'm finding cyclic dependencies
that are too difficult to unravel.  Package A depends on
`openbmc/obmc-pal.h` which is provided by libpal, but libpal depends on
package A.  The obmc-pal has served to break this cycle in the past, so
I am reverting back to this for libpal_0.2.  obmc-pal will install
critical header files and libpal will install the full shared library.

Test Plan: Built a few systems (sp, tp, emeraldpools)

Reviewed By: benwei13

fbshipit-source-id: 0f0a396aa6