fbep: Add server presence pin (#1727)
Summary:
DVT baseboard added presence pin to detect if connected with server.
If server is present, the 12V of emeraldpools had to be enabled by CPLD to sync the power sequence with server
otherwise, BMC had to enable it by itself.
Pull Request resolved: https://github.com/facebookexternal/openbmc.quanta/pull/1727

Test Plan: Build fbep pass and test on DVT baseboard by MFG

Reviewed By: benwei13

fbshipit-source-id: d69138b8ad