ptest-runner: Automatically add recipe to image
Summary:
Not all recipes will be built into the qemux86 image.
Instead of having to constantly add each recipe to the
image, this script just adds it automatically before
building it.

Test Plan:
./tests2/experimental/ptest-runner log-util-v2
(log-util-v2 is not part of the qemux86 image)
Ensure that we get the test output.

./tests2/experimental/ptest-runner libmisc-utils
(Ensure that we get the misutils output)

./tests2/experimental/ptest-runner log-util-v2
(Ensure we get test output of log-util)

./tests2/experimental/ptest-runner log-util-v2 libmisc-utils
(Ensure we get both output).

Reviewed By: benwei13

fbshipit-source-id: 5e0ff5a730