Backport DHCPv6.SendVendorOption from systemd as of 2020-07-06
Summary:
[putting this up for comment mostly because it depends on other stuff that's yet to be landed]

When moving to systemd, the intent is to use systemd-networkd and its DHCPv6 client.

We need to send model name and serial number with DHCPv6 requests from the BMC but the in-tree systemd lacks this capability:

https://www.internalfb.com/intern/wiki/OpenBMC/DHCP_Request_Information/

To help with this, backport the DHCPv6 vendor option capability from systemd as of 2020-07-06 to rocko (systemd 234), warrior (systemd 241) and zeus (systemd 243.2).  This is based off earlier work we fed into systemd upstream (https://github.com/systemd/systemd/pull/15663).

Also, pull in a couple of small fixes for IPv6 with systemd 241 (otherwise it throws errors during interface configuration).

Test Plan:
Successfully build a wedge100 image with all of Luis' systemd changes plus this change.  Also:

For each of the 3 different versions of systemd, configure a test Linux system with the new option:

```
# grep SendVendorOption /etc/systemd/network/eno1.network
SendVendorOption=40981:1:string:hello world
```
On another system, configure router advertisements.  Start tcpdump and boot the test Linux system and wait for DHCPv6 packets.

```
# tcpdump -n -i re0 -s 1500 -w mycap.cap host ether fc:4d:d4:3d:38:94 and port dhcpv6-client
```

Check for the vendor option in the output:

{F243645470}

TODO: configure and test this on a wedge100

fbshipit-source-id: a395985c28