fw-util: Return after unit tests pass
Summary:
Instead of going through parsing the input, just return early.
This prevents the executable failing and ptest thinking
the unit-test failed.

Plus, the original reasoning was to be able to run fw-util on
devserver, but now we have a better "contained" way of running
it on the devserver, so the point of letting the code run
through to the rest of main does not exist.

Test Plan:
devvm350:openbmc :-) (fix-fw-util-unit-test) # ./tests2/experimental/ptest-runner fw-util
Found 437 recipes of which 8 define unit-tests
<snip>
START: ptest-runner
2020-07-15T10:56
BEGIN: /usr/lib/fw-util/ptest
[==========] Running 7 tests from 2 test cases.
[----------] Global test environment set-up.
[----------] 3 tests from ComponentTest
[ RUN      ] ComponentTest.FruListTest
[       OK ] ComponentTest.FruListTest (0 ms)
[ RUN      ] ComponentTest.BaseClassTest
[       OK ] ComponentTest.BaseClassTest (0 ms)
[ RUN      ] ComponentTest.AliasTest
[       OK ] ComponentTest.AliasTest (0 ms)
[----------] 3 tests from ComponentTest (0 ms total)

[----------] 4 tests from BmcComponentTest
[ RUN      ] BmcComponentTest.DefaultSystemVersion
[       OK ] BmcComponentTest.DefaultSystemVersion (0 ms)
[ RUN      ] BmcComponentTest.MTDVersion
[       OK ] BmcComponentTest.MTDVersion (7 ms)
[ RUN      ] BmcComponentTest.MTDFlash
[       OK ] BmcComponentTest.MTDFlash (6 ms)
[ RUN      ] BmcComponentTest.MTDOffsetFlash
[       OK ] BmcComponentTest.MTDOffsetFlash (7 ms)
[----------] 4 tests from BmcComponentTest (20 ms total)

[----------] Global test environment tear-down
[==========] 7 tests from 2 test cases ran. (20 ms total)
[  PASSED  ] 7 tests.
DURATION: 1
END: /usr/lib/fw-util/ptest
2020-07-15T10:56
STOP: ptest-runner

Reviewed By: sammyblueeyes

fbshipit-source-id: 13f658459d