ptest-runner: Run tests of all supported tests as default
Summary:
If user does not pass a recipe to test, then run all
the tests (meeting the behavior of the real ptest-runner).
But since there could be a lot of packages with ptests,
ensure that only the ptests of openBMC recipes are run.

Test Plan: Run `ptest-runner` and ensure it runs all the recipes.

Reviewed By: williamspatrick

fbshipit-source-id: 7ed7f25812