packagegroup-obmc-emmc: add emmc-ext4 feature support
Summary:
We are going to default emmc to use btrfs for better data integrity
support.  Add a new feature 'emmc-ext4' which will be used for systems
already in production that might now want to move up to btrfs (yet).

Test Plan: Built SP with both feature settings and booted on QEMU.

Reviewed By: benwei13

fbshipit-source-id: df49e9fd9d