fby3: Temporary revert EMMC FS from btrfs to ext4
Summary:
During large file transfers, BMC free memory
can be exhausted. The current workaround in
healthd to release kernel caches do not seem to
work. While that is being debugged, this is a
temporary workaround to move back to EXT4.

This diff must be reverted once the original root-cause
is fixed.

Test Plan: Build on fby3 and ensure that the mount scripts point to ext4.

Reviewed By: zhdaniel12

fbshipit-source-id: dd014602cd