emeraldpools: Set the primary network interface as usb0
Summary:
On Emeraldpools, the primary interface to public
is usb0 and not eth0. This change sets hostname
scripts to query for current hostname on that interface.

Test Plan:
Flash on emeraldpools, delete /mnt/data/hostname and reboot.
After boot check if the hostname is correctly updated and `ps` does not show hostname.sh (It terminates, does not linger forever).

Reviewed By: zhdaniel12

fbshipit-source-id: d116259211