CPLD update IOCTL helper (#51)
Summary:
Implement new CPLD update helper to use JTAG IOCTLs.
The two supported IOCTLs are RUN_CYCLE and IOCBITBANG.
RUN_CYCLE is available in 5.3 and BITBANG is available
in 5.6.

This helper is an alternative to
the existing ast_jtag helper which sets JTAG pins one
at a time using sysfs files. The IOCTL method produces
a big improvement in programming times.

The existing cpldhelper interfaces are read/write pin. A
new interface, do_io, is introduced to allow a helper to
read/write the pins in a single operation.

Testing performed:
This change was tested together with additional changes
and showed improvements in JTAG programming times.

In isolation, testing performed was to build a yamp
image and verify the new libcpldupdate_dll_ioctl.so library
was present.

Pull Request resolved: https://github.com/facebookexternal/openbmc.arista/pull/51

Reviewed By: mikechoifb

fbshipit-source-id: 8afafcc389