flashy: Add timeout to WriteFile procedure
Summary:
On memory-stressed devices, WriteFile may cause the process
to freeze (as it waits for the device to sync).

This happened on a yv1 device.

Add a timeout for the write, which will likely fall back to a SafeToReboot
error for the device to reboot and clear up more memory.

Test Plan: Unit tests (`go test ./...`)

Reviewed By: kawmarco

fbshipit-source-id: a8317d160f