ELBERT: Add hack for mgmt linkup on affected protos (#65)
Summary:
ELBERT: Add hack for mgmt linkup on affected protos

There is a rework required to fix certain HW issue on early protos. In order to not completely drop support, there is this script to check and use a hack to mitigate issue temporarily. Long term this script should be removed.

Testing:
Tested by changing SN list and confirming this takes effect.

Pull Request resolved: https://github.com/facebookexternal/openbmc.arista/pull/65

Reviewed By: mikechoifb

fbshipit-source-id: dfd70a6690