Make syslog restart aware of systemd
Summary:
This daemon is on init.d, and its restart sequence is
different from those supervised by runit. I don't think it's worth
abstracting this further.

Test Plan:
Booted a wedge100 with systemd and flashed it with
flashy. It has applied the new image, as desired. The journal logs all
flashy messages, until the very end:

```
 -device mtd:flash0][1666]: 2020/08/26 10:43:21 *** PASSED: Validation with image format 'fit' ***
 Aug 26 10:43:21 fboss96703260-oob.snc1.facebook.com flash_wedge100[1666]: [flashy|/opt/flashy/flash_procedure/flash_wedge100 -imagepath image -device mtd:flash0][1666]: 2020/08/26 10:43:21 Flash device validation passed
 Aug 26 10:43:21 fboss96703260-oob.snc1.facebook.com flash_wedge100[1666]: [flashy|/opt/flashy/flash_procedure/flash_wedge100 -imagepath image -device mtd:flash0][1666]: 2020/08/26 10:43:21 Finished: flash_procedure/flash_wedge100
 ```

And rsyslog is alive at the end of the flash process:

```
[[0;1;[[0m rsyslog.service - System Logging Service
   Loaded: loaded (/lib/systemd/system/rsyslog.service; enabled; vendor preset: enabled)
      Active: [[0;1;32mactive (running)[[0m since Wed 2020-08-26 10:35:03 PDT; 53min ago
```

Reviewed By: vineelasmile, lhl2617

fbshipit-source-id: 1e8c4d224d