Yamp: Avoid writing to FPGA registers after the FPGA has been programmed (#75)
Summary:
The fpga_util.sh script has logic to reset registers back to their
original values after an FPGA upgrade has completed. However, in
cases where the registers being reset are on the same FPGA that
is being programmed, an error may be printed because, depending
on the timing, the FPGA may not be accessible. For example:
/usr/local/bin/fpga_util.sh: line 29: echo: write error: No such device or address

The solution is to avoid writing to registers on the FPGA that has
just been programmed. As part of the programming, the register will
be set to the default value anyway.

Testing performed:
1. Performed an upgrade of the SUP FPGA and ensured that no
   write errors were printed after the upgrade completed.

2. Performed an upgrade of the SCD FPGA and ensured that no
   write errors were printed after the upgrade completed.

Pull Request resolved: https://github.com/facebookexternal/openbmc.arista/pull/75

Test Plan: Done at OEM site.

Reviewed By: tao-ren

fbshipit-source-id: a89b3a3525