mac-util: Add machineoverrides to specify the EEPROM
Summary:
Instead of hard-coding the machine specific EEPROM path/offset
move it to a machine specific header which can be provide
as machineoverrides in the recipe.

Advantage of this method over doing a bbappend in the machine
layer is that with this, you have all the code available in the
recipe and most importantly, multiple machines can share code.

Test Plan: Build mac-util on emeraldpools

Reviewed By: tao-ren

fbshipit-source-id: 8b5067e106