flashy: Add RO offset argument to FlashCp implementation
Summary:
It would rid the need of having to patch the image file in vboot cases
This offset essentially would be a region to skip in both
the image and the flash device.

Simplify also erase, write and verification logic to verify whole active
region instead of block-by-block. This is faster, and makes implementing
the offset logic easier.

Also using a direct Pwrite syscall to write to the flash device instead of wrappers
that may unnecessarily wear off the flash device.

Test Plan:
Unit tests (`go test ./...`)
Manual tests: see next diff.

Reviewed By: kawmarco

fbshipit-source-id: ab21752d28