Move systemd settings to rackmond to common
Summary:
This will be shared with Wedge400 and Wedge40. Our old
special case didn't make sense.

Test Plan:
Built and booted a wedge100 image with it with no
problems. Built a wedge-40 image (the easiest one to build), and it
installs the exact same rackmond files in the exact same places.

Reviewed By: williamspatrick, doranand

fbshipit-source-id: dd80fcecb4