Adapt to systemd
Summary: Add a service for starting mTerm on angelslanding

Test Plan:
If I enable systemd, it installs the unit in the right
place. I am not converting angelslanding to systemd yet, so I haven't
tested if this unit actually works.

The original runit service remains if I don't enable systemd

Reviewed By: doranand

fbshipit-source-id: 922bb80c52