Adapt recipe to systemd
Summary:
Add and systemd services and install them if we are using
such feature.

Please note: there is no inittab in systemd. If we want to replace
`init 0` by a reboot, the equivalent is to replace the
`poweroff.target` unit with `reboot.target`.

Test Plan:
It builds both with systemd (tested on fbal) and with
sysv. No changes on any platforms yet, since I am not converting any
platforms just yet.

Reviewed By: doranand

fbshipit-source-id: 79e5ead851