baudrate algorithm for rackmond
Summary:
## Background

The new G2 PSUs currently in the EVT and DVT phases (for the vendors Delta and Artesyn respectively) provide a new piece of functionality: they support faster communication baud rates.

This new feature allows a non-trivial efficiency gain -- at the highest available baud rate, firmware updates are 20% faster, and overall CPU usage on the BMC is decreased by 50%.

Rackmond needs to be able to support communicating with both V2 (older generation) and G2 (newer generation) PSUs, sometimes in the same rack. However, we only want to raise the baud rate if every unit in the rack can support it.

## The Algorithm
- Every time we try to communicate with a PSU, check whether every unit in the rack can support a higher baud rate.
- If so, check that this unit's baud rate is set to the desired value.
- If it's not, send a command to set it to the desired value.

There are some additional considerations that this diff addresses:

## Graceful Exit

When rackmond starts up, it assumes that all connected units are at the default (lowest) baud rate. We need to intercept any signals that could kill the rackmond process (SIGINT and SIGTERM for now) and when that happens, reset the baud rate for all connected units to the default value.

## Back Off

After N consecutive communication failures (currently N = 10) with a PSU, rackmond needs to back off for a while. The G2 PSUs have an escape hatch, where if we don't send any communication whatsoever to the units for 10 minutes, then they will revert to the default baud rate. This backoff needs to apply to external modbuscmd commands as well, such as the ones that Dynamo uses.

## Interface Differences

This diff also adds a new failure code, 6, to return to clients when they request communication with a unit currently in the 10-minute back-off mode. In addition, it adds the current baud rate to the display in rackmonstatus. There may eventually be a document that we need to add to dozuki somewhere since SiteOps engineers rely on this output.

Test Plan:
The only test plan available is to build rackmond, scp it to a host, and run some manual tests.

Here are my test cases:
- normal operation on a G2 rack (rackmonstatus, modbuscmd reads)
- normal operation on a V2 rack (rackmonstatus, modbuscmd reads)
- Ctrl-C and `sv down rackmond` exits gracefully
- `sv restart rackmond` exits gracefully and starts back up without a hitch
- `modbuscmd a40600a30002` to set the unit to an inaccessible baud rate, wait ten minutes, and make sure `modbuscmd a40300a30001` returns a valid response

**G2 test rack:** rsw1bu.23.snc1
(PSU a6 in this rack is not responding; unrelated to the changes here.)

{F337419294}

**V2 test rack:** rsw1af-oob.03.vll1
{F337231701}

I did not test on any systemd hosts (the new system, instead of sv), but it shouldn't make a difference since both send SIGTERM when you restart the process.

I'd like to be able to test someone literally pulling out the unit and reseating it, ensuring that the device comes back online after 10 minutes. We need to do this in both G2 and V2 racks to ensure that we haven't broken the repair flow in any way. I'll coordinate with Cameron to make sure that gets done as part of the G2 PSU NPI validation during DVT.

Reviewed By: rojer

fbshipit-source-id: 3dbdb397b3