wedge100: disable systemd
Summary:
As an alternative to D24155768 (and based on that diff):

- disable systemd on wedge100 by changing DISTRO_FEATURES
- undo bitbake changes that should be conditional on DISTRO_FEATURES.

The idea is to allow push to continue without systemd if needed and also allow development on systemd to continue.

Test Plan: Built wedge100 with bitbake and flashed fboss8610645-oob.snc1.  It boots multiuser and at first glance looks okay.

Reviewed By: deathowl, tao-ren

fbshipit-source-id: 6061955423