Fix bug in the config
Summary:
fscd was maxing out at 40% and our vendor told us that must be
fixed. I have discussed this also internally with thermal and this is
not a desired behavior and they also confirm it must be fixed. This diff
fix the issue.

Test Plan:
we first run get_fan_speed to take a look at the speed. Then
we change one of them and it works as expected. Since fscd is still on,
fscd brings it down to what is was supposed to be. commands details
output can be found in that paste P144897253

Reviewed By: mikechoifb

fbshipit-source-id: 39d3f2f6a9