rocko: meson: fix bbclass for cross-compiling
Summary:
In the Rocko distribution, the meson.bbclass comes from meta-oe, but in
later distributions it is included in poky.  The version from meta-oe
has issues with cross-compiling that cause meson to generate an ARM
executable with an ELF header directed to use the x86_64 loader.  This
results in executables that are unable to run.

Copy the meson.bbclass from meta-oe to meta-backports/rocko and fix-up
pieces using warrior's poky meson.bbclass as a reference.  Add a BBPATH
to the meta-backports so that the meta-backports/${DISTRO}/classes can
overrule other layers (and thus the backport version is chosen over
meta-oe).

Test Plan:
Run Tiogapass image under QEMU.
Before...
```
   root@bmc-oob:~# kv --help
   -bash: /usr/bin/kv: No such file or directory
   root@bmc-oob:~# strings /usr/bin/kv | head -n1
   /workdir/local/builds/build-minipack/tmp/sysroots-uninative/x86_64-linux/lib/ld-linux-x86-64.so.2
```
After...
```
   root@bmc-oob:~# kv --help
   kv <op> [args]+
```

Reviewed By: tao-ren

fbshipit-source-id: 5a66749426