gpio-ctrl: ignore strncpy warning
Summary:
Newer versions of GCC warn on a strncpy due to possible truncation
of the source string.  In this case, the truncation is acceptable
and rearranging to code to avoid the warning would cause unnecessary
functional changes.  Suppress the warning with pragmas for simplicity.

Test Plan: Build a Zeus-based platform.

Reviewed By: tao-ren

fbshipit-source-id: 76154d405c