Use DISTRO_FEATURES in psmuxmon recipe
Summary:
Undoing D24160378 but still keeping all of
the systemd 'logic' behind DISTRO_FEATURES guard.

This is part of a series of moving systemd-relevant recipes to be
conditional on the DISTRO_FEATURES flag.

Test Plan:
bitbake + flashed on fboss8382003-oob.snc1

https://www.internalfb.com/intern/paste/P146970035/

The device boots and seems to start services

https://www.internalfb.com/intern/paste/P146970046/

Image running:
  root@fboss8382003-oob:~# cat /etc/issue
  OpenBMC Release wedge100-25a7f7c2c2-dirty

Reviewed By: doranand

fbshipit-source-id: 0dc11029bb