Make ipmi recipe work under systemd
Summary:
Wedge* uses ipmi 1.0, which wasn't converted to a systemd
unit and thus we weren't starting before.

Test Plan:
Built with systemd enabled, the unit is present, and there
is a symlink to it from the `multi-user.target.wants` directory.

Built it without systemd, and the sv service is in the right place.

Also, CIT confirms no regressions (I hope...)

Reviewed By: doranand

fbshipit-source-id: cf9ff250a7