fbep: Add lock while setting power capping (#2105)
Summary:
As title

Pull Request resolved: https://github.com/facebookexternal/openbmc.quanta/pull/2105

Test Plan:
Buiid pass and test on fbep (with 2 Nvidia OAMs)

// Enable persistence mode in Host OS
root@localhost ~]# nvidia-smi -pm 1
Enabled persistence mode for GPU 00000000:8A:00.0.
Enabled persistence mode for GPU 00000000:8C:00.0.
All done.

// Set power limit by Emeraldpools BMC
root@bmc-oob:~# asic-util --set-power 300
Set power limit of GPU on slot 0 to 300 Watts
ASIC on slot 1 is not present
ASIC on slot 2 is not present
Set power limit of GPU on slot 3 to 300 Watts
ASIC on slot 4 is not present
ASIC on slot 5 is not present
ASIC on slot 6 is not present
ASIC on slot 7 is not present

Reviewed By: benwei13

fbshipit-source-id: f64557c9f6