rocko: meson: support native building with bbclass
Summary:
When I originally created the backport for the meson.bbclass to support
cross-compiling, I skipped over allowing native building.  Make minor
changes to the bbclass to hopefully support native building.

Confirmed `kv` executable still builds with the target ld-linux.so
instead of the native (which was the original problem).

Test Plan:
```
    $ bitbake kv -c cleanall && bitbake kv
    $ file tmp/work/armv6-fb-linux-gnueabi/libkv/0.1-r1/build/kv

    tmp/work/armv6-fb-linux-gnueabi/libkv/0.1-r1/build/kv: ELF 32-bit LSB executable, ARM, EABI5 version 1 (SYSV), dynamically linked, interpreter /lib/ld-linux.so.3, for GNU/Linux 3.2.0, BuildID[sha1]=6e718b1c5bb06fe72a2154730227eefed52c17fe, with debug_info, not stripped
```

Reviewed By: GoldenBug

fbshipit-source-id: 40c748a353