openbmc-init-build-env: allow symlink'd meta
Summary:
When renaming a meta-layer or machine it is sometimes useful to
be able to leave a symbolic link to the old name so that certain
build tooling continues functioning as-is.  Search for files and
links when looking for the machine.conf file.

Also update similar functionality in netcastle script.

Test Plan: `openbmc-init-build-env fbnd` with Northdome rename.

Reviewed By: benwei13

fbshipit-source-id: 91f571ce87