common: rackmon: ensure a single instance of rackmon-config.py
Summary:
We've seen a few BMCs with multiple instances of "rackmon-config.py",
and this is usually caused by rackmond being restarted by "sv" in a
loop. BMC becomes extremely slow in this case because most CPU resources
are used by rackmon-config.py.

There is no doubt we need to fix rackmond crash loop, but it's still
good to prevent multiple instances of rackmon-config.py.

Test Plan:
1) booted up wedge400: rackmond is launched and configured properly.
2) Launched 2 instances of rackmon-config.py: the 2nd instance failed with
"another instance is running".

Reviewed By: benwei13

fbshipit-source-id: 7387e9795f