Mask named on systemd
Summary:
We don't use any stub DNS resolvers, so this daemon is
useless. Masking it saves us a little bit of boot time, but more
importantly, makes the boot logs less confusing.

Test Plan:
Built with and without systemd. With systemd, the unit is
symlinked to `/dev/null`, which is what masking means. Without it, the
unit just builds normally.

Reviewed By: williamspatrick

fbshipit-source-id: ff96eae163