fby3: GPv3: enable VR fault monitor (#1451)
Summary:
-GPv3 BIC monitored VR P3V3_STBY1/2/3 and P1V8 when it's ready.
If BMC tried to modify the setting of the VRs, the error bit of the VRs will be held.
As a result, GPv3 BIC issues a false alarm event to BMC.

Now, GPv3 BIC provideds a command to let BMC control VR fault monitor.

Pull Request resolved: https://github.com/facebookexternal/openbmc.wiwynn/pull/1451

Test Plan: -build code(pass)

Reviewed By: tao-ren, GoldenBug

fbshipit-source-id: e7cbc8166b