common: libipmb: dump a log in case of truncated resp_buf
Summary:
resp_len is set to MAX_IPMB_RES_LEN (1024) and passed to ipc_send_req(),
and the value will be updated to the actual transfer length when
ipc_send_req() returns. The actual length could be as many as 1024
bytes, which means the value might be truncated when it's assigned to
*res_len (unsigned char type).

Ideally we should update lib_ipmb_handle() interface (extending res_len
from unsigned char to size_t), but it involves a lot of changes (at
caller side). For now, let's just log a warning to make debug easier.

Test Plan:
1) BIOS upgrade succeeded on wedge400.
2) ondiff CIT looks good (no CMM cit results, but CMM is not using the lib).

Reviewed By: binhuang00

fbshipit-source-id: ad0fa7088e