obmc-pal: Set default yes to pal_is_fru_prsnt()
Summary:
This allows platforms to defined a MAX_NUM_FRUS
statically but allow certain FRUs to not be present.

This change makes the default always return True
for the query on whether the FRU is present thus making
certain static platforms simple and not require them
to define this function only to always return true!

Test Plan: Build all platforms

Reviewed By: zhdaniel12

fbshipit-source-id: 1a55d40be0