Replaced calls to Popen(cat|echo|wdtcli) with native calls
Summary:
Some small optimisations in fscd code
 * Changed calls to `Popen("echo > ...)` and `Popen("cat ...")` to native file operations
 * Changed calls to `wdtcli` to native `libwatchdog` calls

Most of the load is still in calls to `Popen("sensor-util ...")`, but changing that requires a bit more work to make a library out of `sensord` code.

Test Plan: Installed changes to a test minipack, checked log for errors and ensured the watchdog wasn't triggered.

Reviewed By: doranand

fbshipit-source-id: 3174ff4521